carlitos/mcp-servers/browser/requirements.txt
browser-use
langchain-openai
python-dotenv
patchright
playwright
browserbase
